Spliting main file into smaller modules
