+++ src/lib/helpers.js
@@ -1,0 +1,5 @@
+function stringify( name ) {
+  return name.toString().replace( /(^\/|\/$)/g, "" );
+}
+
+export default { stringify };
